+++ lib/ethpy/ethpy/hyperdrive/__init__.py
@@ -9,3 +9,4 @@ from .interface import (
     get_hyperdrive_market,
     get_hyperdrive_pool_info,
 )
+from .receipt_breakdown import ReceiptBreakdown
